proxy/requirements.txt
redis>=5.0
PyJWT>=2.0
Pillow>=9.0
